backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
supabase==2.3.0
python-dotenv==1.0.0
pydantic
httpx==0.24.1